move Peripheral Handler to Proxy
